+++ packages/wrangler/src/deployment-bundle/node-compat.ts
@@ -1,6 +1,6 @@
 import { UserError } from "../errors";
 import { logger } from "../logger";
-import { getNodeCompatMode } from "../node-compat";
+import type { Config } from "../config";
 
 /**
  * Wrangler can provide Node.js compatibility in a number of different modes:
@@ -51,6 +51,13 @@ export function validateNodeCompat({
 	const nodejsCompatV2NotExperimental =
 		compatibilityFlags.includes("nodejs_compat_v2");
 
+	if (nodejsCompatV2) {
+		// strip the "experimental:" prefix because workerd doesn't understand it yet.
+		compatibilityFlags[
+			compatibilityFlags.indexOf("experimental:nodejs_compat_v2")
+		] = "nodejs_compat_v2";
+	}
+
 	if (nodejsCompat && nodejsCompatV2) {
 		throw new UserError(
 			"The `nodejs_compat` and `nodejs_compat_v2` compatibility flags cannot be used in together. Please select just one."
@@ -88,4 +95,45 @@ export function validateNodeCompat({
 	}
 
 	return mode;
+}
+
+export function getNodeCompatMode({
+	compatibility_flags,
+	node_compat,
+}: Pick<Config, "compatibility_flags" | "node_compat">) {
+	const nodejsCompat = compatibility_flags.includes("nodejs_compat");
+	const nodejsCompatV2 = compatibility_flags.includes(
+		"experimental:nodejs_compat_v2"
+	);
+
+	let mode: NodeJSCompatMode;
+	if (nodejsCompatV2) {
+		mode = "v2";
+	} else if (nodejsCompat) {
+		mode = "v1";
+	} else if (node_compat) {
+		mode = "legacy";
+	} else {
+		mode = null;
+	}
+
+	return {
+		legacy: node_compat === true,
+		mode,
+		nodejsCompat,
+		nodejsCompatV2,
+	};
+}
+
+/**
+ * The nodejs_compat_v2 flag currently requires an `experimental:` prefix within Wrangler,
+ * but this needs to be stripped before sending to workerd, since that doesn't know about that.
+ *
+ * TODO: Remove this function when we graduate nodejs_v2 to non-experimental.
+ * See https://jira.cfdata.org/browse/DEVDASH-218
+ */
+export function stripExperimentalPrefixes(
+	compatFlags: string[] | undefined
+): string[] | undefined {
+	return compatFlags?.map((flag) => flag.replace(/^experimental:/, ""));
 }
